% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%.IN.\%}
\alias{\%.IN.\%}
\title{Membership Operator}
\usage{
x \%.IN.\% table
}
\arguments{
\item{x}{A R vector or \code{NULL}. See \code{\link{\%in\%}}. The value to
look for.}

\item{table}{A R vector or \code{NULL}. See \code{\link{\%in\%}}. Where to look
for \code{x}.}
}
\value{
A logical vector the same length as \code{x}, or \code{FALSE} when \code{x} is \code{NULL}.
}
\description{
This infix operator is inspired by how the \verb{in} operator works in Python.
}
\details{
Except when \code{x} is \code{NULL}, this operator is the same as \code{\link{\%in\%}}.
Otherwise, when \code{x} is \code{NULL}, it is treated as an empty set, which means
that \code{NULL \%.IN.\% table} is always \code{FALSE}, regardless of what \code{table} is -
including when \code{table} is \code{NULL}.

No checks on \code{table} are performed; they are left to the \code{\%in\%} operator.
}
\examples{
1:10 `\%.IN.\%` c(1,3,5,9)
NULL `\%.IN.\%` c(1,3,5,9)

}
