% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{op-and-null-lgl}
\alias{op-and-null-lgl}
\alias{\%.&&.\%}
\title{Logical 'And' Operator That Works With \code{NULL}}
\usage{
x \%.&&.\% y
}
\arguments{
\item{x, y}{Any types that \code{&&} accepts, or \code{NULL}.}
}
\value{
A length-one logical vector, including if either \code{x} or \code{y} is \code{NULL}. In the
latter case, \code{FALSE} is returned.
}
\description{
This infix function returns \code{FALSE} when \code{x} is falsey (\code{NULL} or \code{FALSE}).
It will also returns \code{FALSE} if \code{x} is truthy and \code{y} is \code{NULL}.
It otherwise works like the operator \code{&&}. It is inspired by the way the
operator \code{&&} works in Ruby.
}
\details{
No checks are performed; they are left to the \code{&&} operator.
}
\examples{
TRUE \%.&&.\% FALSE # FALSE
NULL \%.&&.\% FALSE # FALSE
TRUE \%.&&.\% NULL # FALSE
NULL \%.&&.\% NULL # FALSE

}
\seealso{
\link{op-and-null-nil}
}
