% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{op-and-null-nil}
\alias{op-and-null-nil}
\alias{\%.AND.\%}
\title{Logical 'And' Operator That Works With \code{NULL}}
\usage{
x \%.AND.\% y
}
\arguments{
\item{x, y}{Any types that \code{&&} accepts, or \code{NULL}.}
}
\value{
A length-one logical vector, except if either \code{x} or \code{y} is \code{NULL}. In the
latter case, \code{NULL} is returned.
}
\description{
This infix function returns \code{x} when \code{x} is falsey (\code{FALSE} or \code{NULL}). So it
would return \code{NULL} if \code{x} is \code{NULL}. It otherwise works like the operator
\code{&&}. It is inspired by the way the operator \code{and} works in Ruby.
}
\details{
No checks are performed; they are left to the \code{&&} operator.
}
\examples{
TRUE \%.AND.\% FALSE # FALSE
NULL \%.AND.\% FALSE # NULL
TRUE \%.AND.\% NULL # NULL
NULL \%.AND.\% NULL # NULL

}
\seealso{
\link{op-and-null-lgl}
}
